# Benchmark and improve tree performance - Closes #5349 #5491  