--- a/elements/lisk-tree/README.md
+++ b/elements/lisk-tree/README.md
@@ -8,6 +8,23 @@
 $ npm install --save @liskhq/lisk-tree
 ```
 
+## Benchmarking
+
+Benchmark results for main lisk-tree functions:
+
+|           Function            | 1000 leaves | 10,000 leaves | 100,000 leaves |
+| :---------------------------: | :---------: | :-----------: | :------------: |
+|             build             |    45ms     |     240ms     |     2348ms     |
+|            append             |     3ms     |      3ms      |      4ms       |
+|    generateProof (1 query)    |     5ms     |      5ms      |      6ms       |
+|  generateProof (100 queries)  |    50ms     |     125ms     |     172ms      |
+| generateProof (1000 queries)  |    56ms     |     114ms     |     168ms      |
+| generateProof (10000 queries) |     n/a     |    7993ms     |    14504ms     |
+|     verifyProof (1 query)     |     3ms     |      5ms      |      5ms       |
+|    verifyProof (100 query)    |    45ms     |     106ms     |     166ms      |
+|   verifyProof (1000 query)    |    539ms    |     958ms     |     1592ms     |
+|   verifyProof (10000 query)   |     n/a     |    8632ms     |    15909ms     |
+
 ## License
 
 Copyright 2016-2020 Lisk Foundation