--- a/elements/lisk-tree/test/merkle_tree.spec.ts
+++ b/elements/lisk-tree/test/merkle_tree.spec.ts
@@ -15,8 +15,7 @@ import * as fixture from './fixtures/transaction_merkle_root/transaction_merkle_
 import { EMPTY_HASH, MerkleTree } from '../src/index';
 import { verifyProof } from '../src/verify_proof';
 
-/* eslint-disable jest/no-disabled-tests */
-describe.skip('MerkleTree', () => {
+describe('MerkleTree', () => {
 	describe('constructor', () => {
 		for (const test of fixture.testCases) {
 			describe(test.description, () => {