--- a/framework/test/integration/specs/controller/child_process_channel.spec.ts
+++ b/framework/test/integration/specs/controller/child_process_channel.spec.ts
@@ -66,8 +66,8 @@ const beta = {
 		},
 	},
 };
-
-describe('ChildProcessChannel', () => {
+/* eslint-disable jest/no-disabled-tests */
+describe.skip('ChildProcessChannel', () => {
 	describe('after registering itself to the bus', () => {
 		let alphaChannel: ChildProcessChannel;
 		let betaChannel: ChildProcessChannel;