… protocol-specs/generator_outputs/transaction_merkle_root/transaction_merkle_root.json