### What was the problem?

Benchmark and improve merkle tree performance.

This PR resolves #5349 #5491 

### How was it solved?

Performance was improved by replacing `getTreeStructure` functionality with `locationToHash` map which reduces proof generation to `O(log n)`.

The key to location to hash map is the binary representation of the node index.

**### Benchmark results** 

|    Function                    |      1000 leaves         |        10,000 leaves     |    100,000 leaves        |
| :-----------------------:      | :----------------------: | :----------------------: | :----------------------: |
|         build                  |         45ms             |          240ms           |           2348ms         |
|         append                 |          3ms             |           3ms            |             4ms          |
|  generateProof (1 query)       |          5ms             |           5ms            |             6ms          |
|  generateProof (100 queries)   |          50ms            |           125ms          |            172ms         |
|  generateProof (1000 queries)  |          56ms            |           114ms          |            168ms         |
|  generateProof (10000 queries) |          n/a             |           7993ms         |          14504ms         |
|  verifyProof (1 query)         |          3ms             |           5ms            |             5ms          |
|  verifyProof (100 query)       |         45ms             |           106ms          |            166ms         |
|  verifyProof (1000 query)      |         539ms            |           958ms          |            1592ms        |
|  verifyProof (10000 query)     |          n/a             |           8632ms         |           15909ms        |


Node version used: v12.14.1. Computer Spec: SSD, 6 Core, 16 GB RAM. No special configuration for Node.

### How was it tested?

Benchmark performance using trees with increased `dataLength`.