@shuse2 2020-07-02
it's probably nice to keep the environment information for the benchmark