@mitsuaki-u 2020-07-03
> @mitsuaki-u why we are not including the benchmark details to README?

Further optimization might change benchmark figures, but I can add for now and update it later.